:bug: id에 int들을 long으로 변경

+++ server/src/main/java/com/classic/imteller/api/controller/GameController.java
@@ -37,7 +37,7 @@ public class GameController {
 
     @GetMapping("/rooms/{roomId}")
     @ApiOperation(value = "해당 게임방 정보 받기", notes = "해당 게임방의 정보를 JSON형태로 반환")
-    public ResponseEntity<GameRoomDto> getRoom(@PathVariable final int roomId){
+    public ResponseEntity<GameRoomDto> getRoom(@PathVariable final long roomId){
         // 소켓 or Redis 에서 해당 roomId 정보 가져와서 던져주기
         GameRoomDto gameRoom = new GameRoomDto();
         return new ResponseEntity<GameRoomDto>(gameRoom, HttpStatus.OK);
@@ -45,7 +45,7 @@ public class GameController {
 
     @PostMapping("/rooms/{roomId}/join")
     @ApiOperation(value = "해당 게임방 접속", notes = "해당 게임의 방이름과 비밀번호로 접근")
-    public ResponseEntity<Boolean> joinRoom(@PathVariable final int roomId, @RequestBody final JoinReqDto roomInfo){
+    public ResponseEntity<Boolean> joinRoom(@PathVariable final long roomId, @RequestBody final JoinReqDto roomInfo){
         // 비밀번호 맞는지 여부를 파악하고 boolean값을 보냄
         return new ResponseEntity<Boolean>(true, HttpStatus.OK);
     }
